pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Shubham Patil",email = "shubhampatilsd@gmail.com"},
    {name = "Ishaan Awasthi",email = "ishaanawasthi05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "chromadb (>=0.6.3,<0.7.0)",
    "pillow (>=10.4.0,<12.0.0)",
    "transformers (>=4.21.0,<4.48.0)",
    "torch (>=2.5.1,<3.0.0)",
    "customtkinter (>=5.2.2,<6.0.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "embed-anything (>=0.5.0,<0.6.0)",
    "einops (>=0.8.0,<0.9.0)",
    "pyvips (>=2.2.3,<3.0.0)",
    "pyvips-binary (>=8.16.0,<9.0.0)",
    "moondream (==0.0.5)",
]

[tool.poetry]
packages = [
    { include = "api" }
]

[tool.poetry.scripts]
start = "api.run:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
